implemented search bar and filter

+++ src/main/java/swf/army/mil/brightforgewidget/DTOs/ImageDTO.java
@@ -1,7 +1,7 @@
 package swf.army.mil.brightforgewidget.DTOs;
 
 public record ImageDTO(
-        Integer imageID,
+        Long imageID,
         String imgUrl
 ) {
 }
